# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nepal Summer League Finland 2026 Tournament API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

  # ── Matches ───────────────────────────────────────────────────────────
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List all matches
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, finished]
      responses:
        "200":
          description: Matches list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a match fixture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
...
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match by ID with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match with events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "404":
          description: Not found
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update match (score, status, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}/start:
    post:
      operationId: startMatch
      tags: [matches]
      summary: Start a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}/finish:
    post:
      operationId: finishMatch
      tags: [matches]
      summary: Finish a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}/reset:
    post:
      operationId: resetMatch
      tags: [matches]
      summary: Reset a match back to upcoming
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"

  # ── Goals ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, shortName, primaryColor]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    TeamInput:
      type: object
      required: [name, shortName, primaryColor]
      properties:
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        logoUrl:
          type: ["string", "null"]
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        logoUrl:
          type: ["string", "null"]

    # ── Match ─────────────────────────────────────────────────────────
    Match:
      type: object
      required: [id, homeTeamId, awayTeamId, scheduledTime, pitch, matchNumber, status]
      properties:
        id:
          type: integer
        matchNumber:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homeTeamName:
          type: ["string", "null"]
        awayTeamName:
          type: ["string", "null"]
        homeTeamShortName:
          type: ["string", "null"]
        awayTeamShortName:
          type: ["string", "null"]
        homeScore:
          type: integer
        awayScore:
          type: integer
        scheduledTime:
          type: string
        pitch:
          type: integer
        status:
          type: string
          enum: [upcoming, live, finished]
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    MatchDetail:
      type: object
      required: [id, homeTeamId, awayTeamId, scheduledTime, pitch, matchNumber, status, goals, cards, events]
      properties:
        id:
          type: integer
        matchNumber:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homeTeamName:
          type: ["string", "null"]
        awayTeamName:
          type: ["string", "null"]
        homeTeamShortName:
          type: ["string", "null"]
        awayTeamShortName:
          type: ["string", "null"]
        homeScore:
          type: integer
        awayScore:
          type: integer
        scheduledTime:
          type: string
        pitch:
          type: integer
        status:
          type: string
          enum: [upcoming, live, finished]
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        goals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        events:
          type: array
          items:
            $ref: "#/components/schemas/MatchEvent"
    MatchInput:
      type: object
      required: [homeTeamId, awayTeamId, scheduledTime, pitch, matchNumber]
      properties:
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        scheduledTime:
          type: string
        pitch:
          type: integer
        matchNumber:
          type: integer
    MatchUpdate:
      type: object
      properties:
        homeScore:
          type: integer
        awayScore:
          type: integer
        scheduledTime:
          type: string
        pitch:
          type: integer
        status:
          type: string
          enum: [upcoming, live, finished]

    # ── Goal ─────────────────────────────────────────────────────────
    Goal:
      type: object
      required: [id, matchId, teamId, minute]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        scorerName:
          type: ["string", "null"]
        assistName:
          type: ["string", "null"]
        minute:
          type: integer
        isOwnGoal:
          type: boolean
        createdAt:
          type: string
...
    Card:
      type: object
      required: [id, matchId, teamId, cardType, minute]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
        cardType:
          type: string
          enum: [yellow, red]
        minute:
          type: integer
        createdAt:
          type: string
...
    MatchEvent:
      type: object
      required: [id, matchId, eventType, minute]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        eventType:
          type: string
          enum: [goal, own_goal, yellow_card, red_card, substitution, kickoff, halftime, fulltime, var_review, injury]
        minute:
          type: integer
        description:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
        createdAt:
          type: string
...