# lib/api-zod/src/generated/types/matchDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nepal Summer League Finland 2026 Tournament API
 * OpenAPI spec version: 0.1.0
 */
import type { Card } from './card';
import type { Goal } from './goal';
import type { MatchDetailStatus } from './matchDetailStatus';
import type { MatchEvent } from './matchEvent';

export interface MatchDetail {
  id: number;
  matchNumber: number;
  homeTeamId: number;
  awayTeamId: number;
  /** @nullable */
  homeTeamName?: string | null;
  /** @nullable */
  awayTeamName?: string | null;
  /** @nullable */
  homeTeamShortName?: string | null;
  /** @nullable */
  awayTeamShortName?: string | null;
  homeScore?: number;
  awayScore?: number;
  scheduledTime: string;
  pitch: number;
  status: MatchDetailStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  createdAt?: string;
  goals: Goal[];
  cards: Card[];
  events: MatchEvent[];
}