# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nepal Summer League Finland 2026 Tournament API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchInput {
  homeTeamId: number;
  awayTeamId: number;
  scheduledTime: string;
  pitch: number;
  matchNumber: number;
}