# lib/api-zod/src/generated/types/matchUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nepal Summer League Finland 2026 Tournament API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchUpdateStatus } from './matchUpdateStatus';

export interface MatchUpdate {
  homeScore?: number;
  awayScore?: number;
  scheduledTime?: string;
  pitch?: number;
  status?: MatchUpdateStatus;
}