# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nepal Summer League Finland 2026 Tournament API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  name?: string;
  shortName?: string;
  primaryColor?: string;
  /** @nullable */
  logoUrl?: string | null;
}